Add job queuing capability to Genie for protection
Add a new job status QUEUED and implement methods to allow jobs
to be stored in the database with a QUEUED status.  Using a
new parameter for max system jobs, we can limit the total number
of running jobs and tailor it to our clusters.

New prop: com.netflix.genie.server.max.system.jobs