Ensure attachments for queued jobs are saved
When we queue a job we need to save its attachments to
S3 because they are not persisted in the database.  Add
a property defining the root location to store these on S3.
Each job will have its own directory.  Allow ability to
delete if it's enabled, but don't cause anything to fail
if it doesn't work since we typically don't allow direct
deletes from S3.

Minor cleanup and bump version for next build.