Breakup finalizeJob for reliability and add process env vars
Break up the finalizeJob method into several transactions to isolate the
different steps from each other and avoid the optimistic locking
exceptions.  First separate job finalization from releasing of queued
jobs.  Next, make a single transactional method to find a queued job
and release it in the same transaction, then add the ability for it to
be retried if it fails with an optimistic locking exceptions.  Move
out attachment restoring and job running from the transaction (job
running is already transactional).

Also add new environment variables to the process for max disk usage and
min memory available to start the process.  This is so we can move the
repeated logic from the setup scripts to the job launcher.