Remove forwarding logic that doesn't work
Job forwarding doesn't work in our environment because it requires
running tons more Netflix components including their discovery server.
As is, this logic is causing random 503s which it shouldn't.
We already have logic for queueing jobs and for checking that we
have memory and disk space available -- so remove this to let
those checks do the right thing.